đẩy tạm lỗi đâu tính sau

+++ backend/src/main/java/com/thubongshop/backend/imports/ImportRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ImportRepository extends JpaRepository<Import, Long> {
+public interface ImportRepository extends JpaRepository<Import, Integer > {
     List<Import> findByStatus(String status);
-    List<Import> findBySupplierId(Long supplierId);
+    List<Import> findBySupplierId(Integer  supplierId);
+    
 }
